Flee Implementation
The NPC will now flee away from the attacker when attacked.  Right now this is just using the walk animation at double speed and triple movement speed.  This can be refined with a unique animation.  It also runs in a straight line.  This should later be refined to pathfinding or dynamically updating lines.